+++ src/main/java/com/besson/arknights/block/custom/columbian/BeverageCabinet.java
@@ -12,10 +12,10 @@ public class BeverageCabinet extends FurnitureHorizontalFacingBlock {
     public BeverageCabinet(Settings settings) {
         super(settings);
     }
-    private static final VoxelShape SHAPE_N = Block.createCuboidShape(0, 0, 4, 16, 16, 16);
-    private static final VoxelShape SHAPE_W = Block.createCuboidShape(4, 0, 0, 16, 16, 16);
-    private static final VoxelShape SHAPE_S = Block.createCuboidShape(0, 0, 0, 16, 16, 12);
-    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, 0, 12, 16, 16);
+    private static final VoxelShape SHAPE_N = Block.createCuboidShape(-1.95, 0, 5, 17.9, 15.75, 16);
+    private static final VoxelShape SHAPE_W = Block.createCuboidShape(5, 0, -1.9, 16, 15.75, 17.95);
+    private static final VoxelShape SHAPE_S = Block.createCuboidShape(-1.9, 0, 0, 17.95, 15.75, 11);
+    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, -1.95, 11, 15.75, 17.9);
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {