+++ src/main/java/com/besson/arknights/block/custom/columbian/BeverageMenu.java
@@ -4,18 +4,28 @@ import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ShapeContext;
+import net.minecraft.util.function.BooleanBiFunction;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 
 public class BeverageMenu extends FurnitureHorizontalFacingBlock {
     public BeverageMenu(Settings settings) {
         super(settings);
     }
-    private static final VoxelShape SHAPE_N = Block.createCuboidShape(-5, 0, 15, 19, 25, 16);
-    private static final VoxelShape SHAPE_W = Block.createCuboidShape(15, 0, -3, 16, 25, 21);
-    private static final VoxelShape SHAPE_S = Block.createCuboidShape(-3, 0, 0, 21, 25, 1);
-    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, -5, 1, 25, 19);
+    private static final VoxelShape SHAPE_N = VoxelShapes.combineAndSimplify(
+            Block.createCuboidShape(0, 13, 15, 19, 25, 16),
+            Block.createCuboidShape(-6, 0, 15, 13, 11, 16), BooleanBiFunction.OR);
+    private static final VoxelShape SHAPE_W = VoxelShapes.combineAndSimplify(
+            Block.createCuboidShape(15, 13, -3, 16, 25, 16),
+            Block.createCuboidShape(15, 0, 3, 16, 11, 22), BooleanBiFunction.OR);
+    private static final VoxelShape SHAPE_S = VoxelShapes.combineAndSimplify(
+            Block.createCuboidShape(-3, 13, 0, 16, 25, 1),
+            Block.createCuboidShape(3, 0, 0, 22, 11, 1), BooleanBiFunction.OR);
+    private static final VoxelShape SHAPE_E = VoxelShapes.combineAndSimplify(
+            Block.createCuboidShape(0, 13, 0, 1, 25, 19),
+            Block.createCuboidShape(0, 0, -6, 1, 11, 13), BooleanBiFunction.OR);
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {