+++ src/main/java/com/besson/arknights/block/custom/columbian/BolivarianCoffeeBean.java
@@ -12,10 +12,10 @@ public class BolivarianCoffeeBean extends FurnitureHorizontalFacingBlock {
     public BolivarianCoffeeBean(Settings settings) {
         super(settings);
     }
-    private static final VoxelShape SHAPE_N = Block.createCuboidShape(0, 0, 14, 16, 24, 16);
-    private static final VoxelShape SHAPE_W = Block.createCuboidShape(14, 0, 0, 16, 24, 16);
-    private static final VoxelShape SHAPE_S = Block.createCuboidShape(0, 0, 0, 16, 24, 2);
-    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, 0, 2, 24, 16);
+    private static final VoxelShape SHAPE_N = Block.createCuboidShape(0, 0, 13, 16, 24, 16);
+    private static final VoxelShape SHAPE_W = Block.createCuboidShape(13, 0, 0, 16, 24, 16);
+    private static final VoxelShape SHAPE_S = Block.createCuboidShape(0, 0, 0, 16, 24, 3);
+    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, 0, 3, 24, 16);
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {