+++ src/main/java/com/besson/arknights/block/custom/columbian/DiningChair.java
@@ -27,7 +27,7 @@ public class DiningChair extends FurnitureHorizontalFacingBlock {
     @Override
     public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
         if (!world.isClient){
-            return SeatEntity.create(world,pos,0.35,player,state.get(FACING));
+            return SeatEntity.create(world,pos,0.3,player,state.get(FACING));
         }
         return ActionResult.SUCCESS;
     }